# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: XBoard
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
...
paths:
...
  '/api/v1/guest/payment/notify/{method}/{uuid}':
    get:
      summary: '支付网关异步回调通知 (Webhook)'
      operationId: Webhook
      description: "用于接收第三方支付平台异步发送的订单支付结果通知，验证签名后更改内部订单状态并下发服务。\n该接口无需身份认证，完全依赖网关签名校验。"
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: plain/text
      tags:
        - '用户端 API (User & Guest)'
      security: []
    parameters:
      -
        in: path
        name: method
        description: ''
        example: natus
        required: true
        schema:
          type: string
      -
        in: path
        name: uuid
        description: ''
        example: 145d25e7-6a57-3fa8-b247-30ebcaf7daa8
        required: true
        schema:
          type: string
...